test/requirements.txt
setuptools<58.0.0
cup==3.2.31
nose-progressive==1.5.2
nose~=1.3.7
parameterized
pymysql
nose_xunitmp
flaky==3.7.0
fuzzywuzzy
python-Levenshtein
PyGithub